First commit.
This is a skeleton site. Much more work needs to be done. It is a simple stylesheet/site, and has support for columns if we want them through col_33/50/75 style markup.

Much more work is necessary for fancy bits (auto-generation of tables based on assignments, etc.), but it is enough to get us started.